Update to gridstack.js
There is an issue with the gridstack when placed inside a single page application. The gridstack is still busy updating "data-gs-style-id" even though there is no gridstack element available in the whole HTML. This cause performance issues in the application because the gridstack is busy updating "data-gs-style-id" even though there is no any gridstack instance is present. However, this won't happen if I refresh the page. But, in a single page application, chances are very less that we refresh the whole page. 

A simple check to see whether is there any gridstack instance available or not will prevent this issue from happening. The proposed solution will only restrict "GridStack.prototype._initStyles " and "GridStack.prototype._updateStyles" to proceed further if there's gridstack is present. 

There might be some other alternative methods available but I would like to propose this change or seek more advice on this.

Many thanks,